fixed reconnection flow
