+++ server/crud/user/basic.js
@@ -7,5 +7,8 @@ module.exports = {
 async function get_user(socket_id) {
     return await db.collection('user').findOne({
         socket_id: socket_id
-    }, {lat: 1, long: 1, socket_id: 1, energy: 1, posture: 1, angle: 1});
+    }, {
+        lat: 1, long: 1, socket_id: 1, energy: 1,
+        posture: 1, angle: 1, last_read_patch_notes: 1
+    });
 }
