+++ server/crud/user/user.js
@@ -1,5 +1,6 @@
 const db = require('../db/db').get_db();
 const crud_look_around = require('../interact/look_around');
+const crud_connection = require('./connection');
 
 module.exports = {
     create_user,
@@ -49,12 +50,7 @@ async function create_user(email, socket) {
             if(error.code === 11000) {
                 // duplicate email / returning user
                 socket.send({data: "Welcome back!"});
-                // update socket id
-                db.collection('user').updateOne({
-                    email: email
-                }, {
-                    $set: {socket_id: socket.id}
-                });
+                crud_connection.add_connection(email, socket.id);
             } else {
                 console.error(error);
             }
